+++ src/main/java/com/buuz135/industrial/item/BookManualItem.java
@@ -7,7 +7,10 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResult;
+import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
 
 public class BookManualItem extends IFCustomItem {
@@ -18,8 +21,15 @@ public class BookManualItem extends IFCustomItem {
 
     @Override
     public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
-        playerIn.openGui(IndustrialForegoing.instance, GuiHandler.BOOK, worldIn, playerIn.getPosition().getX(), playerIn.getPosition().getY(), playerIn.getPosition().getZ());
-        return super.onItemRightClick(worldIn, playerIn, handIn);
+        BlockPos pos = playerIn.getPosition();
+        if (playerIn.isSneaking()) {
+            RayTraceResult result = rayTrace(worldIn, playerIn, false);
+            if (result != null && result.typeOfHit == RayTraceResult.Type.BLOCK) {
+                pos = result.getBlockPos();
+            }
+        }
+        playerIn.openGui(IndustrialForegoing.instance, GuiHandler.BOOK, worldIn, pos.getX(), pos.getY(), pos.getZ());
+        return new ActionResult<>(EnumActionResult.SUCCESS, playerIn.getHeldItem(handIn));
     }