+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Changed `IPlantRecollectable` to `PlantRecollectable` and changed to a IForgeRegistry )
 + Added Protein Reactor with JEI Handler and Craftweaker support.
 + Added Protein Generator that produces 320RF/t with protein.
++ Sneak+Right Click a IF Block with the manual in the hand will open the manual entry of the block directly
 
 # 1.4.3
 + Fixed [#65](https://github.com/Buuz135/Industrial-Foregoing/issues/65)